api/requirements.txt
spacy==2.3.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
Flask==1.1.2
Flask-Cors==3.0.9
google-cloud-language==1.3.0
requests==2.24.0
sumy==0.8.1
textblob==0.15.3
vaderSentiment==3.3.2
newspaper3k==0.2.8